Discord challenge accepted. I regret my choices.
It should be now easier to follow the logic and find errors. I did not alter the functions beyond efficiency.
Some IF trees were prunned of irrelevant steps or compresed, a couple of switch block were introduced where it made sense, a simple loop added. All non-IF tree modifications include comments as to what I'm doing and how.

There are many things that could be done to this file, but I don't use autotrimps so I just corrected the most obvious offenders.