Feat: Automatically Use Bone Charges When Maxed
This adds a new feature that can be turned on allowing automatic use of bone charges when they are maxed to prevent them from being wasted.

It supports both Helium (U1) and Radon (U2) universes, has an option to toggle it on just for dailies, and you are also able to specify the start zone. If no zone is specified it will automatically calculate 10% of your highest zone cleared and start using bone charges from that zone onwards.